Clear command + clean-up

+++ bot/cogs/admin.py
@@ -28,16 +28,23 @@ class Admin(commands.Cog):
 
     @commands.command()
     @commands.has_any_role("Moderator", "Guides of the Void")
-    async def say(self, ctx, *args):
+    async def say(self, ctx, channel_name, *message):
         await ctx.message.delete()
-        channel = discord.utils.get(ctx.guild.text_channels, name=args[0])
-        msg = " ".join(args[1:])
-        if channel is not None:
-            await channel.send(msg)
+        channel = discord.utils.get(ctx.guild.text_channels, name=channel_name)
+        if channel:
+            if message:
+                msg = " ".join(message)
+                await channel.send(msg)
+            else:
+                raise commands.errors.BadArgument("Please enter a message")
+        else:
+            raise commands.errors.BadArgument(
+                "Please enter a valid channel, no leading #"
+            )
 
     @commands.command()
     @commands.has_any_role("Moderator", "Guides of the Void")
-    async def punish(self, ctx, arg):
+    async def punish(self, ctx, user_mention):
         await ctx.message.delete()
         guild = ctx.guild
         ch_list = [
@@ -51,17 +58,19 @@ class Admin(commands.Cog):
 
         for ch in ch_list:
             channel = discord.utils.get(guild.text_channels, name=ch)
-            ping = await channel.send(arg)
-            await ping.delete()
-            await asyncio.sleep(2)
+            if channel:
+                ping = await channel.send(user_mention)
+                await ping.delete()
+                await asyncio.sleep(2)
 
         await asyncio.sleep(45)
 
         for ch in ch_list:
             channel = discord.utils.get(guild.text_channels, name=ch)
-            ping = await channel.send(arg)
-            await ping.delete()
-            await asyncio.sleep(2)
+            if channel:
+                ping = await channel.send(user_mention)
+                await ping.delete()
+                await asyncio.sleep(2)
 
     @commands.command()
     async def horny(self, ctx):
@@ -88,6 +97,28 @@ class Admin(commands.Cog):
     async def react(self, ctx, msg_id, emoji):
         msg = await ctx.fetch_message(msg_id)
         await msg.add_reaction(emoji)
+
+    @commands.command()
+    async def cringe(self, ctx):
+        await ctx.send(
+            "https://cdn.discordapp.com/attachments/441286267548729345/709160523907989524/EVACI9dUcAQp2Mb.png"
+        )
+
+    @commands.command(aliases=["clean"])
+    @commands.has_any_role("Moderator", "Guides of the Void")
+    async def clear(self, ctx, amount):
+        await ctx.message.delete()
+        deleted = 0
+        async for msg in ctx.channel.history(limit=20):
+            await msg.delete()
+            deleted += 1
+            if deleted == int(amount):
+                break
+        reply = await ctx.send(
+            f":white_check_mark: Deleted the last {deleted} messages! :white_check_mark:"
+        )
+        await asyncio.sleep(3)
+        await reply.delete()
 
 
 def setup(bot):
